+++ src/main/kotlin/step2/contorller/View.kt
@@ -3,5 +3,5 @@ package step2.contorller
 interface View {
     fun getMathExpression(): List<String>?
 
-    fun printConclusion(result: Result<Int>)
+    fun printConclusion(result: Result<Double>)
 }
